moved styles into their own folders
